Issue #4220: Modified AvoidNestedBlocksCheckTest.java and moved its input files to the avoidnestedblocks subdirectory
